First Draft of moving archival node into harmony
